# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Payfluence API",
    "version": "v1.0.0",
    "description": "\nPayfluencer is a creator monetization platform that connects brands with content creators through performance-based bounties. This API enables both creators and companies to interact with the platform programmatically and manage workflows including bounty listing, content submission, review, and communication.\n\nCore Capabilities:\n- Creator authentication and profile management\n- Company bounty creation with detailed targeting criteria\n- Submission of creator content for review and reward tracking\n- Real-time chat between creators, clients, and platform admins\n- Status tracking for content submissions (Draft, Submitted, Accepted, Paid)\n\nAuthentication:\nAll endpoints require authentication using Supabase JWT tokens. Only Gmail-based authentication is currently supported.\n\nContent Guidelines:\nBounty listings may include content restrictions such as language, age rating, platform, and visibility requirements.\n\nFor support or questions, contact the Payfluencer team at support@payfluencer.xyz.\n",
    "license": {
      "name": "Copyright 2025 Payfluence"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8001"
    }
  ],
  "paths": {
...
    "/api/v1/report": {
      "post": {
        "summary": "Create a new report",
        "tags": [
          "Report"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id",
                  "company_id",
                  "title",
                  "description",
                  "platform"
                ],
                "properties": {
                  "user_id": {
                    "type": "string",
                    "description": "User ID"
                  },
                  "company_id": {
                    "type": "string",
                    "description": "Company ID"
                  },
                  "title": {
                    "type": "string",
                    "description": "Report title"
                  },
                  "description": {
                    "type": "string",
                    "description": "Report description"
                  },
                  "platform": {
                    "type": "string",
                    "description": "Platform name"
                  },
                  "status": {
                    "type": "string",
                    "description": "Report status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Report created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "report": {
                          "$ref": "#/components/schemas/Report"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get report by ID",
        "tags": [
          "Report"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Report ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Report found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Report found"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "report": {
                          "$ref": "#/components/schemas/Report"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Report not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update a report (user can update their own, except status)",
        "tags": [
          "Report"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "user_id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "Report ID"
                  },
                  "user_id": {
                    "type": "string",
                    "description": "User ID"
                  },
                  "title": {
                    "type": "string",
                    "description": "Report title"
                  },
                  "description": {
                    "type": "string",
                    "description": "Report description"
                  },
                  "platform": {
                    "type": "string",
                    "description": "Platform name"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Report updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "report": {
                          "$ref": "#/components/schemas/Report"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a report (user can delete their own)",
        "tags": [
          "Report"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "user_id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "Report ID"
                  },
                  "user_id": {
                    "type": "string",
                    "description": "User ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Report deleted successfully"
                    },
                    "data": {
                      "type": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/report/status": {
      "put": {
        "summary": "Update report status (company only)",
        "tags": [
          "Report"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "company_id",
                  "status"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "Report ID"
                  },
                  "company_id": {
                    "type": "string",
                    "description": "Company ID"
                  },
                  "status": {
                    "type": "string",
                    "description": "New status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Report status updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "report": {
                          "$ref": "#/components/schemas/Report"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Report": {
        "type": "object",
        "required": [
          "user_id",
          "company_id",
          "title",
          "description",
          "platform"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Report ID"
          },
          "user_id": {
            "type": "string",
            "description": "User ID who created the report"
          },
          "company_id": {
            "type": "string",
            "description": "Company ID the report is for"
          },
          "title": {
            "type": "string",
            "description": "Report title"
          },
          "description": {
            "type": "string",
            "description": "Report description"
          },
          "platform": {
            "type": "string",
            "description": "Platform name"
          },
          "status": {
            "type": "string",
            "description": "Report status (draft, submitted, approved, rejected)"
          },
          "createdAt": {
            "type": "string",
            "description": "Report creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "Report last update timestamp"
          }
        }
      },
...
    }
  },
...
}